Movies-react.js-project
This is a React.js and Redux movie app that fetches movie data from an API and displays a list of movies with their posters. Clicking on a movie shows detailed information like its description and larger poster. The app uses Redux for state management and is designed to be responsive.